pyproject.toml
[tool.poetry]
name = "crypto-python"
version = "0.1.0"
description = ""
authors = ["Yip Jun Kai <yipjunkai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.19.0"

[tool.poetry.dev-dependencies]
black = "^23.10.1"
tqdm = "^4.66.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
